Refactor file paths and remove unused components
This commit updates the README and sample files to replace hardcoded user paths with environment variable references, enhancing portability. Additionally, it removes the `config.rs`, `highlight.rs`, `models.rs`, and `session.rs` files, streamlining the codebase by eliminating unused components. The changes improve the overall structure and maintainability of the project.
