+++ src/lib.rs
@@ -1,7 +1,9 @@
+pub mod api;
 pub mod cli;
 pub mod config;
-pub mod highlight;
+pub mod error;
 pub mod local_tools;
 pub mod models;
 pub mod session;
+pub mod ui;
