+++ src/local_tools/builtins/mod.rs
@@ -1,0 +1,4 @@
+mod read_file;
+
+pub use read_file::handle_read_file;
+
