+++ src/local_tools/tools.rs
@@ -1,8 +1,6 @@
-use super::paths::safe_resolve_path;
-use super::registry::{LocalSettings, LocalToolRegistry};
+use super::registry::LocalToolRegistry;
 use colored::Colorize;
 use serde_json::{json, Value};
-use std::fs;
 
 pub fn format_tools_for_llm(registry: &LocalToolRegistry) -> Vec<Value> {
     registry
@@ -52,68 +50,4 @@ pub async fn call_local_tool(
     // Call the handler (now async)
     let handler = &tool.handler;
     handler(arguments, registry.settings()).await
-}
-
-// Tool handlers
-
-pub fn handle_read_file(args: &Value, settings: &LocalSettings) -> Result<String, String> {
-    let path_str = args
-        .get("path")
-        .and_then(|v| v.as_str())
-        .ok_or_else(|| "Missing required argument: path".to_string())?;
-
-    if settings.verbose {
-        eprintln!(
-            "{}",
-            format!(
-                "[tools] Resolving path: '{}' (base_dir={})",
-                path_str,
-                settings.base_dir.display()
-            )
-            .dimmed()
-        );
-    }
-
-    // Safely resolve the path
-    let resolved_path = safe_resolve_path(path_str, &settings.base_dir)?;
-
-    if settings.verbose {
-        eprintln!(
-            "{}",
-            format!("[tools] Resolved path: {} -> {}", path_str, resolved_path.display())
-                .dimmed()
-        );
-    }
-
-    // Check if file exists
-    if !resolved_path.exists() {
-        return Err(format!("File not found: {}", path_str));
-    }
-
-    // Check if it's a file (not a directory)
-    if !resolved_path.is_file() {
-        return Err(format!("Path is not a file: {}", path_str));
-    }
-
-    // Check file size
-    let metadata =
-        fs::metadata(&resolved_path).map_err(|e| format!("Failed to read file metadata: {}", e))?;
-    if metadata.len() > settings.max_file_size_bytes {
-        return Err(format!(
-            "File too large: {} bytes (max: {} bytes)",
-            metadata.len(),
-            settings.max_file_size_bytes
-        ));
-    }
-
-    if settings.verbose {
-        eprintln!(
-            "{}",
-            format!("[tools] Reading file: {} ({} bytes)", resolved_path.display(), metadata.len())
-                .dimmed()
-        );
-    }
-
-    // Read file (UTF-8 only)
-    fs::read_to_string(&resolved_path).map_err(|e| format!("Failed to read file: {}", e))
 }
