--- src/models.rs
@@ -1,91 +1,0 @@
-use serde::{Deserialize, Serialize};
-use serde_json::Value;
-
-#[derive(Serialize, Deserialize, Clone)]
-pub struct Message {
-    pub role: String,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub content: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub tool_calls: Option<Vec<ToolCall>>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub tool_call_id: Option<String>,
-}
-
-#[derive(Serialize, Deserialize, Clone)]
-pub struct ToolCall {
-    pub id: String,
-    #[serde(rename = "type")]
-    pub tool_type: String,
-    pub function: FunctionCall,
-}
-
-#[derive(Serialize, Deserialize, Clone)]
-pub struct FunctionCall {
-    pub name: String,
-    pub arguments: String,
-}
-
-#[derive(Serialize, Deserialize, Clone)]
-pub struct Session {
-    pub session_id: String,
-    pub last_updated: chrono::DateTime<chrono::Local>,
-    pub messages: Vec<Message>,
-}
-
-#[derive(Serialize, Clone)]
-pub struct Reasoning {
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub effort: Option<String>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub max_tokens: Option<u32>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub exclude: Option<bool>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub enabled: Option<bool>,
-}
-
-#[derive(Serialize)]
-pub struct RequestBody {
-    pub model: String,
-    pub messages: Vec<Message>,
-    pub stream: bool,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub reasoning: Option<Reasoning>,
-    #[serde(skip_serializing_if = "Option::is_none")]
-    pub tools: Option<Vec<Value>>,
-}
-
-#[derive(Deserialize)]
-pub struct Citation {
-    pub url: String,
-    pub title: String,
-    #[allow(dead_code)]
-    pub content: Option<String>,
-}
-
-#[derive(Deserialize)]
-pub struct Annotation {
-    #[serde(rename = "type")]
-    pub annotation_type: String,
-    pub url_citation: Option<Citation>,
-}
-
-#[derive(Deserialize)]
-pub struct Delta {
-    pub content: Option<String>,
-    pub annotations: Option<Vec<Annotation>>,
-    pub reasoning: Option<String>,
-    #[allow(dead_code)]
-    pub tool_calls: Option<Vec<ToolCall>>,
-}
-
-#[derive(Deserialize)]
-pub struct Choice {
-    pub delta: Option<Delta>,
-}
-
-#[derive(Deserialize)]
-pub struct StreamResponse {
-    pub choices: Option<Vec<Choice>>,
-}
