+++ src/ui/mod.rs
@@ -1,0 +1,5 @@
+pub mod highlight;
+pub mod output;
+
+pub use output::{display_content, display_reasoning, display_tool_error, display_tool_result};
+
